reorganize + rename tests / dirs
