+++ convokit/tests/test_utils.py
@@ -266,5 +266,6 @@ def reload_corpus_in_db_mode(corpus):
     corpus_id = uuid4().hex
     corpus.dump(corpus_id, base_path=".")
     db_corpus = Corpus(corpus_id, storage_type="db")
+    print(f"corpus id is {corpus_id}")
     shutil.rmtree(corpus_id)
     return db_corpus
